Renamespace SizeInterface
