+++ src/RequestField/Field/Size.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace App\RequestField\Field;
 
-use App\ErrorResponse\SizeInterface;
+use App\RequestField\SizeInterface;
 
 readonly class Size implements SizeInterface
 {
