+++ src/RequestField/ScalarRequirementsInterface.php
@@ -4,8 +4,6 @@ declare(strict_types=1);
 
 namespace App\RequestField;
 
-use App\ErrorResponse\SizeInterface;
-
 interface ScalarRequirementsInterface extends RequirementsInterface
 {
     /**
